Install Clojure Web App on Vagrant (#6)
* After bringing over plexus.clojure-web-app from plexus/ansible_playbooks.

* Removed lein from tasks.

Lein is already installed earlier.
Lein "self-install" is deprecated.

* Use python3 for ansible.

It will be the default in the next Ansible version.
Note that python3 is already installed by default on Ubuntu 18.04.
Note that the default for Ansible is still python2.

* Use python3 for Ansible

* Added aptitude to installed packages.

Mainly to shut up a warning from Ansible 2.8 when doing a
dist-upgrade.

Sorted the list of packages.

* Moved dist-upgrade to after the add/remove of packages.

More appropriate location.

* Explicitly get the boolean value of a bare var.

To handle a deprecation warning.

* Need to install the python3 version of psycopg2.

Probably not the best way of doing things. Likely should use a conditional.

* Don't update cache after installing new packages.

Since this will be done later with a dist-upgrade.

* Added some TODOs to make the datomic role idempotent later.

* Copied over plexus.clojurians-log role.

* Explicitly specify the primary group for the clojure_app user.

* Updated Datomic install.

Ensure that the config is owned by the clojure_app user.

Updated the checkout of the logs to use current Ansible practices.

Updated the repo to point to the ddemo data. Will need to fix this
later.

* Made the datomic install idempotent and usable in check mode.

* Bring back the update_cache: yes for apt packages.

Removing it did not remove the warning the first time.

* Added authorized_keys for lispyclouds, victorb and bombaywalla.

* Updated the pre-reveive hook for git 2.14+

See https://github.com/dokku/dokku/issues/2796 for details.
I wanted to make minimal changes to get things running.
Not sure this is a good long-term solution, but should work for now.

* Added documentation to deploy the clojurians-log app.

Currently to a vagrant instance.